Functionality Achieved
The core functionality of the systems is full swing. Many bugs to fix including, import bug if data is not accurate on yahoo finance, bug relating to importing stocks at 3pm, and many more. Dead code everywhere next step is to clean up the execution of the code before we work on a better indicator. Fully functioning SMA indicator however.